refactor: criando hooks para separação de responsabilidades
