+++ src/hooks/use-products/index.tsx
@@ -1,0 +1,25 @@
+import { useEffect, useState } from "react";
+import type { Product } from "../../pages/products/types";
+import { getProducts } from "../../services/product-service";
+
+export const useProducts = () => {
+    const [products, setProducts] = useState<Product[]>([]);
+    const [isLoading, setIsLoading] = useState<boolean>(true);
+
+    useEffect(() => {
+        const fetchProducts = async () => {
+            try {
+                const productList = await getProducts();
+                setProducts(productList);
+            } catch (error) {
+                console.error("Erro ao buscar produtos:", error);
+            } finally {
+                setIsLoading(false);
+            }
+        }
+
+        fetchProducts();
+    }, []);
+
+    return { products, isLoading };
+}
